feat:comment modify

+++ src/pages/Detail/components/CommentItem/index.tsx
@@ -1,25 +1,37 @@
 import {
-  Avatar, Box, Paper, IconButton, Rating, Stack, Typography, CircularProgress,
+  Avatar, Box, Paper, IconButton, Rating, Stack, Typography,
+  CircularProgress,
 } from '@mui/material';
+import { useConfirm } from 'material-ui-confirm';
 import React from 'react';
 import { CommentDataType } from 'types/data';
 import DeleteOutlineIcon from '@mui/icons-material/DeleteOutline';
 import { useDeleteActivityCommentMutation } from 'store/activity/activityService';
+import EditIcon from '@mui/icons-material/Edit';
 import { toast } from 'react-toastify';
 import { useAppSelector } from 'store';
 import { selectUserInfo } from 'store/user/userSlice';
+import { format, parseISO } from 'date-fns';
+
+interface CommentItemType extends CommentDataType {
+  onOpen: () => void
+}
 
 function CommentItem({
-  username, content, userAvatar, rate, id, userId,
-}: CommentDataType) {
+  username, content, userAvatar, rate, id, userId, modifiedAt, onOpen,
+}: CommentItemType) {
   const [deleteComment, { isLoading: isDeleting }] = useDeleteActivityCommentMutation();
   const userInfo = useAppSelector(selectUserInfo);
+  const confirm = useConfirm();
 
   const handleDelete = () => {
-    deleteComment(id)
-      .unwrap()
-      .then(() => toast.success('成功刪除留言'))
-      .catch((err: any) => toast.error(err.data.message));
+    confirm({ title: '確定刪除此留言?' })
+      .then(() => {
+        deleteComment(id)
+          .unwrap()
+          .then(() => toast.success('成功刪除留言'))
+          .catch((err: any) => toast.error(err.data.message));
+      });
   };
 
   return (
@@ -28,15 +40,24 @@ function CommentItem({
         <Avatar src={userAvatar} />
         <Stack sx={{ flexGrow: 1 }}>
           <Typography variant="body1">{username}</Typography>
-          <Typography variant="caption">2023/5/29</Typography>
+          <Typography variant="caption">
+            上次編輯時間:
+            {' '}
+            {format(parseISO(modifiedAt), 'yyyy/MM/dd')}
+          </Typography>
         </Stack>
         {(userInfo && userInfo.id === userId)
         && (
-          <IconButton onClick={handleDelete}>
-            {isDeleting
-              ? <CircularProgress size="1em" />
-              : <DeleteOutlineIcon />}
-          </IconButton>
+          <>
+            <IconButton onClick={onOpen}>
+              <EditIcon />
+            </IconButton>
+            <IconButton onClick={handleDelete}>
+              {isDeleting
+                ? <CircularProgress size="1em" />
+                : <DeleteOutlineIcon />}
+            </IconButton>
+          </>
         )}
       </Box>
       <Rating
@@ -47,6 +68,7 @@ function CommentItem({
       <Typography variant="body2">
         {content}
       </Typography>
+
     </Paper>
   );
 }
